requirements.txt
fastapi
uvicorn
langchain
langchain-community
pyspellchecker
deep-translator
